--config is only for 'gwas.py gwas' and 'gwas.py pgrs'
Previously gwas.py --config was available also for merge-plink2, merge-regenie and merge-saige.
However the actual config is not used by those three operations.
A complex issue here is that ``gwas.py gwas`` and ``gwas.py pgrs`` commands are executed by a user
interactively from a terminal (potentially not via container, or via container but with different set of --bind and/or --home argument).
``gwas.py merge-XXX`` are executed from a SLURM script via containers, with --bind and/or --home arguments set according to config.yaml.
Because of this difference in singularity --bind / --home, it's likely that ``gwas.py merge-XXX`` won't have access to config.yaml located in some custom folder (e.g. user home).
A simple fix is to not use --config in ``gwas.py merge-XXX`` because it's not needed.